requirements.txt
future==0.18.2
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pkg-resources==0.0.0
torch==1.5.0
torchvision==0.6.0
